Fix for #993 (Analysis does not produce any results).
Have made error reporting (via popup dialog) and error logging more
"consistent". (For Post Processing, a Floor Mask is not mandatory -
but limits the results produced. Both a Floor Plan and a Mask are
needed for (sensible) Results Analysis so that has been made clearer.

Also included is replacement of the (corner) help button with a save
button. This is only visible when changes to the workbench have been
made - it makes it clear (and easier) to the user that a save is needed.